backend/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
bcrypt
python-jose
pydantic
email-validator